TEST : 템플릿 메서드 패턴을 이용해 Controller 중복 코드 제거
