+++ src/main/java/com/example/bookreservationserver/user/controller/UserController.java
@@ -5,8 +5,8 @@ import com.example.bookreservationserver.user.dto.JoinRequest;
 import com.example.bookreservationserver.user.dto.UserResponse;
 import com.example.bookreservationserver.user.service.AuthService;
 import com.example.bookreservationserver.user.service.JoinService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,12 +15,11 @@ import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 
 @RestController
+@RequiredArgsConstructor
 @Slf4j
 public class UserController {
-    @Autowired
-    private JoinService joinService;
-    @Autowired
-    private AuthService authService;
+    private final JoinService joinService;
+    private final AuthService authService;
 
     @PostMapping("/api/user/join")
     public UserResponse join(@RequestBody @Valid JoinRequest joinRequest){
