+++ src/test/java/com/example/bookreservationserver/book/controller/BookControllerTest.java
@@ -5,25 +5,18 @@ import com.example.bookreservationserver.book.dto.BookRequestDto;
 import com.example.bookreservationserver.book.service.BookAddService;
 import com.example.bookreservationserver.book.service.BookDeleteService;
 import com.example.bookreservationserver.book.service.BookSearchService;
-import com.google.gson.*;
+import com.example.bookreservationserver.skeleton.ControllerTest;
 import com.google.gson.reflect.TypeToken;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
-import java.lang.reflect.Type;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -32,8 +25,7 @@ import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.doReturn;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@ExtendWith(MockitoExtension.class)
-class BookControllerTest {
+class BookControllerTest extends ControllerTest {
 
     @InjectMocks
     private BookController bookController;
@@ -46,19 +38,6 @@ class BookControllerTest {
 
     @Mock
     private BookSearchService bookSearchService;
-
-    private MockMvc mockMvc;
-
-    private Gson gson;
-
-    @BeforeEach
-    public void setUp(){
-        mockMvc = MockMvcBuilders.standaloneSetup(bookController).build();
-
-        GsonBuilder gsonBuilder = new GsonBuilder();
-        gsonBuilder.registerTypeAdapter(LocalDate.class, new LocalDateSerializer());
-        gson = gsonBuilder.setPrettyPrinting().create();
-    }
 
     @Test
     @DisplayName("책 추가 성공")
@@ -192,14 +171,8 @@ class BookControllerTest {
         );
     }
 
-    // gson 라이브러리 사용시
-    // LocalDate type 컨버팅에 필요
-    private static class LocalDateSerializer implements JsonSerializer<LocalDate>{
-        private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-
-        @Override
-        public JsonElement serialize(LocalDate src, Type typeOfSrc, JsonSerializationContext context) {
-            return new JsonPrimitive(formatter.format(src));
-        }
+    @Override
+    protected Object getController() {
+        return bookController;
     }
 }
