+++ src/test/java/com/example/bookreservationserver/skeleton/ControllerTest.java
@@ -1,0 +1,41 @@
+package com.example.bookreservationserver.skeleton;
+
+import com.google.gson.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+
+import java.lang.reflect.Type;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+@ExtendWith(MockitoExtension.class)
+public abstract class ControllerTest {
+    protected MockMvc mockMvc;
+
+    protected Gson gson;
+
+    @BeforeEach
+    public void setUp(){
+        mockMvc = MockMvcBuilders.standaloneSetup(getController()).build();
+
+        GsonBuilder gsonBuilder = new GsonBuilder();
+        gsonBuilder.registerTypeAdapter(LocalDate.class, new LocalDateSerializer());
+        gson = gsonBuilder.setPrettyPrinting().create();
+    }
+
+    protected abstract Object getController();
+
+    // gson 라이브러리 사용시
+    // LocalDate type 컨버팅에 필요
+    private static class LocalDateSerializer implements JsonSerializer<LocalDate> {
+        private static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+
+        @Override
+        public JsonElement serialize(LocalDate src, Type typeOfSrc, JsonSerializationContext context) {
+            return new JsonPrimitive(formatter.format(src));
+        }
+    }
+}
